requirements.txt
python-telegram-bot
azure-ai
requests
azure-functions
